IN-306 ref: change structure folder
