+++ src/widgets/index.ts
@@ -1,0 +1,2 @@
+export * from './headers'
+export * from './sidebar'
